+++ clang-tools-extra/test/clang-tidy/checkers/bugprone/argument-comment-literals.cpp
@@ -7,7 +7,6 @@
 // RUN:     bugprone-argument-comment.CommentStringLiterals: true, \
 // RUN:     bugprone-argument-comment.CommentNullPtrs: true, \
 // RUN:     bugprone-argument-comment.CommentCharacterLiterals: true}}" --
-
 struct A {
   void foo(bool abc);
   void foo(bool abc, bool cde);